function operators > visitor : renaming a ill-named header file
