+++ src/solver/optim-model-filler/include/antares/solver/optim-model-filler/TimeDependentLinearExpression.h
@@ -21,7 +21,7 @@
 #pragma once
 #include <span>
 
-#include <antares/expressions/visitors/VariadicNodeFunctionVisit.h>
+#include <antares/expressions/visitors/visitFunctionNodeArgs.h>
 
 #include "LinearExpression.h"
